==> Analyses/05.1a.run_gwas_coloc.ipynb <==
setwd("/projects/CARDIPS/analysis/epigenome_resource")
source("analyses/jennifer/notebooks/functions.R")

set.seed(5366)
library(coloc)

## **Prepare input for GWAS**

qtls = fread("analyses/tim/gwas_coloc/scripts/qtls_090424.txt", data.table = T)
qtls$taskid = c(1:nrow(qtls))
qtls$path_exists = file.exists(qtls$path)

table(qtls$path_exists)

head(qtls,2)



table(qtls$path_exists)
qtls$egene = "TRUE"
qtls$analysis = qtls$data_type

## **Launch job**

input_file="/projects/CARDIPS/analysis/epigenome_resource/analyses/tim/gwas_coloc/scripts/qtls_090424.txt"
manifest_file="/projects/CARDIPS/analysis/epigenome_resource/analyses/jennifer/gwas_independent/subset_manifest.txt"
script="path_to_05.1b.run_qtl_gwas_coloc.sh"
outdir="/projects/CARDIPS/analysis/epigenome_resource/analyses/tim/gwas_coloc2"
cmd = paste("qsub -N gwas_coloc2 -t", paste0("1-", nrow(qtls), ":1"), 
            "-tc 300 -pe smp 2 -V -cwd -o coloc_gwas2.out -e coloc_gwas2.err", script, input_file, manifest, outdir)

message(cmd)
# system(cmd)

## **concatenate all GWAS results**

tissues = c("iPSC", "PPC", "CVPC")
analyses = c("eqtls", "caqtls", "haqtls")

coloc_results_dir = "/projects/CARDIPS/analysis/epigenome_resource/analyses/tim/gwas_coloc2/"

table(qtls$tissue, qtls$data_type)

manifest_file = "analyses/tim/gwas_coloc/scripts/manifest_subset.txt"
manifest = fread(manifest, data.table = F)


pipeline = "/projects/CARDIPS/analysis/epigenome_resource/analyses/jennifer/notebooks/08.04.summarize_gwas.R"
manifest_file = "analyses/tim/gwas_coloc/scripts/manifest_subset.txt"

for (t in tissues)
{
    for (a in analyses)
    {
        files = list.files(paste("analyses/tim/gwas_coloc2", a, t, sep = "/"))
        message(paste(a, t, length(files)), appendLF = F)
        if (length(files) > 0)
        {
            cmd = paste("Rscript", pipeline, 
                        "--analysis", a, 
                        "--tissue", t, 
                        "--manifest_file", manifest_file, 
                        "--coloc_results_dir", coloc_results_dir)
            
            log_out = paste(getwd(), "analyses/tim/gwas_coloc2/logs", paste(a, t, "summarize.out", sep = "_"), sep = "/")
            log_err = paste(getwd(), "analyses/tim/gwas_coloc2/logs", paste(a, t, "summarize.err", sep = "_"), sep = "/")
            run_qsub(name = paste(a, t, sep = "_"), cmd = cmd, threads = 4, log_out = log_out, log_err = log_err, exec = T)
        }
    }
}


## **aggregate**

summary = as.data.frame(rbindlist(lapply(tissues, function(t)
{
    as.data.frame(rbindlist(lapply(analyses, function(a)
    {
        file = paste(coloc_results_dir, "results", paste(paste(a, t, "summary", sep = "_"), "txt", sep = "."), sep = "/")
        message(file)
        if (file.exists(file))
        {
            message(paste(t, a),  appendLF = F)
            fread(file, data.table = F)
        } else
        {
            message(paste("Missing", t, a))
        }
    })))
}))) %>% mutate(p.gwas = as.double(p.gwas), p.eqtl = as.double(p.eqtl)) %>% dplyr::rename(type = discovery_order) 

table(summary$analysis, summary$tissue)
